frontend/requirements.txt
# Core UI framework
streamlit

# Data manipulation and numerical operations
pandas
numpy

# --- Google Cloud Libraries ---
google-genai
google-cloud-aiplatform
google-cloud-bigquery
google-cloud-bigquery-storage
google-cloud-storage
google-auth
google-auth-httplib2

db-dtypes



